requirements.txt
python-telegram-bot==20.7
supabase==2.3.0
python-dotenv==1.0.0
APScheduler==3.10.4
openai==1.6.1
requests==2.31.0

